# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "GoWVP Owl 的 HTTP API 文档，包含设备、通道、GB28181、PTZ、录像与系统接口。\n\n鉴权方式：\n大多数管理接口需要在请求头中携带 `Authorization: Bearer \u003ctoken\u003e`。\n可先调用 `/login/key` 获取 RSA 公钥，再调用 `/login` 完成登录。\n\n返回约定：\n1. 成功时通常返回业务对象，或 `{ \"msg\": \"ok\" }`\n2. 失败时通常返回 `{ \"msg\": \"错误原因\" }`\n3. GB28181 查询类接口通常会同时返回结构化结果与原始 XML",
        "title": "GoWVP Owl API",
        "contact": {},
        "version": "0.0.1"
    },
    "basePath": "/",
    "paths": {
...
        "/app/upgrade": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "以 SSE 方式输出下载进度事件，事件类型包含 start、progress、error、complete",
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "System"
                ],
                "summary": "执行应用升级",
                "responses": {
                    "200": {
                        "description": "SSE 事件流",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/app/version/check": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "检查新版本",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.checkVersionOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/configs/info": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Config"
                ],
                "summary": "获取系统配置摘要",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerConfigInfoOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/configs/info/sip": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Config"
                ],
                "summary": "修改 SIP 配置",
                "parameters": [
                    {
                        "description": "SIP 配置",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerSIPConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerMessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/proxy/sms/{path}": {
            "get": {
                "description": "将请求透明转发到配置中的流媒体服务器 HTTP 端口",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MediaServer"
                ],
                "summary": "反向代理流媒体 HTTP 请求",
                "parameters": [
                    {
                        "type": "string",
                        "description": "代理路径",
                        "name": "path",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "代理响应",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.SwaggerConfigInfoOutput": {
            "type": "object",
            "properties": {
                "sip": {
                    "description": "当前系统使用的 SIP 配置",
                    "allOf": [
                        {
                            "$ref": "#/definitions/api.SwaggerSIPConfig"
                        }
                    ]
                }
            }
        },
...
        "api.SwaggerErrorResponse": {
            "type": "object",
            "properties": {
                "msg": {
                    "description": "错误信息",
                    "type": "string",
                    "example": "参数错误"
                }
            }
        },
...
        "api.SwaggerMessageResponse": {
            "type": "object",
            "properties": {
                "msg": {
                    "description": "处理结果说明",
                    "type": "string",
                    "example": "ok"
                }
            }
        },
...
        "api.SwaggerSIPConfig": {
            "type": "object",
            "properties": {
                "domain": {
                    "description": "SIP 域",
                    "type": "string",
                    "example": "3402000000"
                },
                "enable_tls": {
                    "description": "是否启用 SIP-TLS",
                    "type": "boolean",
                    "example": false
                },
                "id": {
                    "description": "平台 20 位国标编码",
                    "type": "string",
                    "example": "34020000002000000001"
                },
                "password": {
                    "description": "注册鉴权密码",
                    "type": "string",
                    "example": "123456"
                },
                "port": {
                    "description": "SIP TCP/UDP 监听端口",
                    "type": "integer",
                    "example": 5060
                },
                "require_message_auth": {
                    "description": "是否要求 MESSAGE/NOTIFY 做 Digest 鉴权",
                    "type": "boolean",
                    "example": false
                },
                "strict_source_check": {
                    "description": "是否严格校验源 IP",
                    "type": "boolean",
                    "example": true
                },
                "tls_cert": {
                    "description": "TLS 证书路径",
                    "type": "string",
                    "example": "configs/certs/sip.crt"
                },
                "tls_key": {
                    "description": "TLS 私钥路径",
                    "type": "string",
                    "example": "configs/certs/sip.key"
                },
                "tls_port": {
                    "description": "SIP-TLS 监听端口",
                    "type": "integer",
                    "example": 5061
                }
            }
        },
...
        "api.checkVersionOutput": {
            "type": "object",
            "properties": {
                "current_version": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "has_new_version": {
                    "type": "boolean"
                },
                "new_version": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}